pyproject.toml
[project]
name = "torchtab"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "chris-santiago", email = "cjsantiago@gatech.edu"},
]
dependencies = [
    "pytorch-tabular>=1.1.0",
    "hydra-core>=1.3.2",
    "wandb>=0.17.8",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.6.3",
]
